Changing some fields to long instead of int as int might be too small. Added some more documentation.

+++ src/nl/ansuz/android/steam/dota/response/match/GetMatchDetailsResponse.java
@@ -112,7 +112,7 @@ public class GetMatchDetailsResponse {
 	 * Numeric match ID.
 	 */
 	@SerializedName("match_id")
-	public int matchId;
+	public long matchId;
 	
 	/**
 	 * A "sequence number", representing the order in which matches were recorded.
